requirements.txt
absl-py==0.7.1
astor==0.8.0
Click==7.0
docutils==0.14
gast==0.2.2
grpcio==1.21.1
h5py==2.9.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
mock==3.0.5
numpy==1.16.4
pandas==0.24.2
protobuf==3.8.0
python-dateutil==2.8.0
python-geohash==0.8.5
pytz==2019.1
six==1.12.0
tensorboard==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
termcolor==1.1.0
Werkzeug==0.15.4
